+++ src/main/java/com/healthcare/system/services/implementation/ComplaintServiceImpl.java
@@ -1,10 +1,15 @@
 package com.healthcare.system.services.implementation;
 
+import com.healthcare.system.entities.Appointment;
 import com.healthcare.system.entities.Complaint;
+import com.healthcare.system.exceptions.ValidationException;
+import com.healthcare.system.exceptions.WrongCredentials;
 import com.healthcare.system.repositories.ComplaintRepository;
 import com.healthcare.system.repositories.implementation.ComplaintRepositoryImpl;
 import com.healthcare.system.services.ComplaintService;
 import java.util.List;
+
+import static com.healthcare.system.util.Verification.verifyCredentials;
 
 public class ComplaintServiceImpl implements ComplaintService {
 
@@ -14,17 +19,24 @@ public class ComplaintServiceImpl implements ComplaintService {
         this.complaintRepository = complaintRepository;
     }
     @Override
-    public Complaint findById(int id) {
+    public Complaint findById(int id) throws WrongCredentials {
+        if(complaintRepository.findById(id) == null) {
+            throw new WrongCredentials("Complaint with id: " + id + " does not exist");
+        }
         return complaintRepository.findById(id);
     }
 
     @Override
-    public void deleteById(int id) {
+    public void deleteById(int id) throws WrongCredentials {
+        if(complaintRepository.findById(id) == null) {
+            throw new WrongCredentials("Complaint with id: " + id + " does not exist");
+        }
         complaintRepository.deleteById(id);
     }
 
     @Override
-    public void update(Complaint complaint) {
+    public void update(Complaint complaint) throws ValidationException {
+        verifyCredentials(Complaint.class,complaint);
         complaintRepository.update(complaint);
     }
 
@@ -34,7 +46,8 @@ public class ComplaintServiceImpl implements ComplaintService {
     }
 
     @Override
-    public void save(Complaint complaint) {
+    public void save(Complaint complaint) throws ValidationException {
+        verifyCredentials(Complaint.class,complaint);
         complaintRepository.save(complaint);
     }
 }
